requirements.txt
aiohttp==2.0.7
appdirs==1.4.3
async-timeout==1.2.1
chardet==3.0.2
multidict==2.1.4
olefile==0.44
packaging==16.8
Pillow==4.1.1
pymongo==3.4.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
yarl==0.10.1
